+++ config.cpp
@@ -33,7 +33,7 @@ class CfgMods {
 		creditsJson = "gebsfish/Scripts/Credits.json";
 		author="Geb";
 		authorID="0";
-		version="3.3";
+		version="3.3.2";
 		extra=0;
 		type="mod";
 		dependencies[]= {